plant_care_assistant/requirements.txt
numpy
pandas
opencv-python
pillow
scikit-learn
Flask
